fixed imports and created paginator function
