+++ cogs/informative.py
@@ -1,10 +1,11 @@
-import discord
-from discord.ext import commands
-import voxelbotutils as vbu
-import utils
 
 import asyncio
 import typing
+
+import discord
+from discord.ext import commands
+import voxelbotutils as vbu
+from cogs import utils
 
 class Informative(vbu.Cog):
 
@@ -146,58 +147,8 @@ class Informative(vbu.Cog):
                 fish_string = [f"**{' '.join(fish_type.split('_')).title()}**" for fish_type, fish_info in fish_types.items()]
                 fields.append((rarity.title(), "\n".join(fish_string)))
 
-            # Create an embed
-            curr_index = 1
-            curr_field = fields[curr_index - 1]
-            embed = utils.create_bucket_embed(ctx.author, curr_field, "**Bestiary**\n")
-
-            fish_message = await ctx.send(embed=embed)
-
-            valid_reactions = ["◀️", "▶️", "⏹️", "🔢"]
-            [await fish_message.add_reaction(reaction) for reaction in valid_reactions]  # Add the pagination reactions to the message
-
-            def reaction_check(reaction, user):
-                return user == ctx.author and str(reaction.emoji) in valid_reactions and reaction.message == fish_message
-
-            while True:  # Keep paginating until the user clicks stop
-                try:
-                    chosen_reaction, _ = await self.bot.wait_for('reaction_add', timeout=60.0, check=reaction_check)
-                    chosen_reaction = chosen_reaction.emoji
-                except asyncio.TimeoutError:
-                    chosen_reaction = "⏹️"
-                
-                index_chooser = {
-                    "◀️": max(1, curr_index - 1),
-                    "▶️": min(len(fields), curr_index + 1)
-                }
-                
-                if chosen_reaction in index_chooser.keys():
-                    curr_index = index_chooser[chosen_reaction]  # Keep the index in bounds
-                    curr_field = fields[curr_index - 1]
-
-                    await fish_message.edit(embed=utils.create_bucket_embed(ctx.author, curr_field, "**Bestiary**\n"))
-
-                elif chosen_reaction == "⏹️":
-                    await fish_message.clear_reactions()
-                    break  # End the while loop
-
-                elif chosen_reaction == "🔢":
-                    number_message = await ctx.send(f"What page would you like to go to? (1-{len(fields)}) ")
-
-                    # Check for custom message
-                    def message_check(message):
-                        return message.author == ctx.author and message.channel == fish_message.channel and message.content.isdigit()
-
-                    user_message = await self.bot.wait_for('message', check=message_check)
-                    user_input = int(user_message.content)
-
-                    curr_index = min(len(fields), max(1, user_input))
-                    curr_field = fields[curr_index - 1]
-
-                    await fish_message.edit(embed=utils.create_bucket_embed(ctx.author, curr_field, "**Bestiary**\n"))
-                    await number_message.delete()
-                    await user_message.delete()
-           
+            utils.paginate(ctx, fields, ctx.author, "**Bestiary**\n")
+        
         else:
             for rarity, fish_types in self.bot.fish.items():
                 for _, fish_info in fish_types.items():
