+++ cogs/shop.py
@@ -5,7 +5,7 @@ import voxelbotutils as vbu
 import discord
 from discord.ext import commands
 
-from cogs import utils
+from cogs from cogs import utils
 
 FISH_SHOP_EMBED = discord.Embed(title="Fish Shop")
 FISH_SHOP_EMBED.add_field(name="Fish Bags", value="These are bags containing a fish of a random rarity", inline=False)
