+++ cogs/utils/misc_utils.py
@@ -1,6 +1,7 @@
 import discord
 import random
 import math
+import asyncio
 
 # Does all the xp stuff
 async def xp_finder_adder(self, user: discord.User, played_with_fish):
@@ -58,12 +59,12 @@ def get_fixed_field(field):
 
     return fixed_field
 
-def create_bucket_embed(user, field):
+def create_bucket_embed(user, field, custom_title=None):
     """
     Creates the embed for the pagination page for the fishbucket
     """
     embed = discord.Embed()  # Create a new embed to edit the message
-    embed.title = f"**{user.display_name}'s Fish Bucket**\n"
+    embed.title = custom_title or f"**{user.display_name}'s Fish Bucket**\n"
     embed.add_field(name=f"__{field[0]}__", value=field[1], inline=False)
     return embed
 
@@ -72,6 +73,59 @@ def create_bucket_embed(user, field):
 #     embed.title = "Commands (anything in quotes is a variable, and the quotes may or may not be needed)"
 #     embed.add_field(name=f"__{field[0]}__", value=field[1], inline=False)
 #     return embed
+
+async def paginate(ctx, fields, user, custom_str=None):
+    bot = ctx.bot
+    curr_index = 1
+    curr_field = fields[curr_index - 1]
+    embed = create_bucket_embed(user, curr_field, custom_str)
+
+    fish_message = await ctx.send(embed=embed)
+
+    valid_reactions = ["◀️", "▶️", "⏹️", "🔢"]
+    [await fish_message.add_reaction(reaction) for reaction in valid_reactions]  # Add the pagination reactions to the message
+
+    def reaction_check(reaction, user):
+        return user == ctx.author and str(reaction.emoji) in valid_reactions and reaction.message == fish_message
+
+    while True:  # Keep paginating until the user clicks stop
+        try:
+            chosen_reaction, _ = await bot.wait_for('reaction_add', timeout=60.0, check=reaction_check)
+            chosen_reaction = chosen_reaction.emoji
+        except asyncio.TimeoutError:
+            chosen_reaction = "⏹️"
+        
+        index_chooser = {
+            "◀️": max(1, curr_index - 1),
+            "▶️": min(len(fields), curr_index + 1)
+        }
+        
+        if chosen_reaction in index_chooser.keys():
+            curr_index = index_chooser[chosen_reaction]  # Keep the index in bounds
+            curr_field = fields[curr_index - 1]
+
+            await fish_message.edit(embed=create_bucket_embed(user, curr_field, custom_str))
+
+        elif chosen_reaction == "⏹️":
+            await fish_message.clear_reactions()
+            break  # End the while loop
+
+        elif chosen_reaction == "🔢":
+            number_message = await ctx.send(f"What page would you like to go to? (1-{len(fields)}) ")
+
+            # Check for custom message
+            def message_check(message):
+                return message.author == ctx.author and message.channel == fish_message.channel and message.content.isdigit()
+
+            user_message = await bot.wait_for('message', check=message_check)
+            user_input = int(user_message.content)
+
+            curr_index = min(len(fields), max(1, user_input))
+            curr_field = fields[curr_index - 1]
+
+            await fish_message.edit(embed=create_bucket_embed(user, curr_field, custom_str))
+            await number_message.delete()
+            await user_message.delete()
         
 def seconds_converter(time):
     if 5_400 > time >= 3_600:
